objDetectMod/requirement-gpu.txt
opencv-python==4.2.0.34
tensorflow-gpu==1.14.0
scipy 
matplotlib
grpcio-tools
pillow